+++ packages/components/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useHash';
 export * from './useMenuItems';
+export * from './useSearch';
 export * from './useStickySidebar';